Use Elixir for bootstrapping
The previous rust-based parser was overkill for the time being, since
setting up the parser was taking too much time, when existing tools like
yacc and bison are easier to set up for parsing.

Since the language is intended to have strong Erlang/Elixir ties, it
seems appropriate to write the bootstrap compiler in it.

Included here is an intial run at tokenizing the source using leex with
some tests added as well.