Remove chars from the lexer and parser
In the future, it might be good to implement ways to get codepoints and
graphemes from a string as opposed to just looking it at as a list of
chars. This approach allows for more natural processing of strings, but
it is also costly. Until then, the char is removed.