Add if-expressions to the bootstrapped kernel
Also added in a macro for true and false, which, like in Elixir, are
just macros that return `:true` and `:false`.

Since `true` and `false` are keywords in Elixir, added in to the ident
parsing is a part that puts an underscore after keywords so that they
can be valid.