Remove spaces from parsing
Spaces were originally added to be able to tell the difference between
macro calls and function calls, though this might not be completely
necessary, or could be handled in the lexer.

The way group (and group_bare) was implemented was made to be more
similar to how it is handled in the Elixir parser.