Begin wrapping Erlang types with `f64`
Erlang doesn't have NaN, inf, -inf, or signed zeros, and so this makes
it difficult down the road for Aero to be consistent across different
platforms.

The aero_core_f64 module wraps the behavior of basic floating point
arithmetic in Erlang with NaN, inf, and -inf. Signed zeros are not
included at the moment since they are rare and require more thought to
implement.