Add `f64` utilities using a Rust-based NIF
This adds the basic structure for creating Rust-based NIFs which the
conversions in place to make them work for Aero.

The bulk of the Rust `f64` functions are now available to be called in
the Erlang `aero_core_f64` module.

The macro for wrapping and exporting NIFs is a little clumsy since the
first array is used for wrapping in Aero types, and the second is use
for exporting those using Rustler. It would be better to do that all
together since we can get the arity inside the macro.

This all of course requires `rustup` to be installed to build.