Add grammar for most parts of the language
This has a lot of stuff in it, and uses operator parsing to do macro
call parsing with fake inserted operators and right precedence. This
should all be enough to parse most things, of course, a lot is subject
to change.

Removed the integer sizes, since this language is targeting BEAM for
now, so the data types should align well to BEAM.

There are a lot of defined operators but after this, many will probably
not be used right away and really are there to have their precedence
reserved.