Add `->>` and `<<-` to lexing/parsing; remove some old ops
`->>` will act as an impure function operator, replacing the `proc`
macro. `<<-` in prefix mode is a receive next operator, and `<<-` in
infix mode is a send operator.

Removed container for binaries `<<>>` to replace with `#bits()` and
`#bytes()` later on, removing spaceship operator `<=>` for now, and
removed `\\` which will be replaced with `=` in loop initialization
and for default function arguments.

Fixed a few assoc and precedence cases.

None of this has been integrated into the Aero.Kernel module, as work
underway for removing the Elixir code generation in favor of compiling
to Core Erlang.